+++ lib/dynamic_theme_picker.dart
@@ -2,4 +2,5 @@
 export 'src/theme_color_picker_widget.dart';
 export 'src/theme_color_picker.dart';
 export 'src/theme_controller.dart';
-export 'src/theme_changer.dart';  // <-- Add this line
+export 'src/theme_changer.dart';
+export 'src/theme_dialog_button.dart';